Add files via upload
Scipy's Lognormal acts weird when data is close to 0.  10^-2 is fine, but 10^-6 is not.
It also goes crazy at random places.  If this happens, rely on loc=0, scale=1 or floc=0.
floc=0 is always wrong for data that has been shifted.  loc=0, scale=1 works for data that is shifted to the origin.